Rename ActivityPool to ActivityService, and move it to core/activity.py
Also reindent the file while at it.
